+++ web/Dockerfile
@@ -3,17 +3,22 @@ FROM php:8.3-apache
 # Set the ServerName to avoid warnings
 RUN echo "ServerName localhost" >> /etc/apache2/apache2.conf
 
-
-# Enable necessary Apache modules
+# Enable Apache rewrite module
 RUN a2enmod rewrite
 
-# Copy application files
+# Set Apache's DocumentRoot to the public folder
+RUN sed -i 's|/var/www/html|/var/www/html/public|g' /etc/apache2/sites-available/000-default.conf \
+    && sed -i 's|<Directory /var/www/html>|<Directory /var/www/html/public>|g' /etc/apache2/apache2.conf
+
+# Copy application files (including public folder)
 COPY . /var/www/html/
 
-# Set permissions
-RUN chown -R www-data:www-data /var/www/html
+# Set correct permissions
+RUN chown -R www-data:www-data /var/www/html \
+    && chmod -R 755 /var/www/html
 
-# Expose port 80 in the container
+# Expose port 80
 EXPOSE 80
 
+# Start Apache in foreground
 CMD ["apache2-foreground"]
