+++ web/app/views/staff/staff.php
@@ -1,0 +1,19 @@
+<?php require_once __DIR__ .'/../../includes/security/session_check.php';
+requireLogin();
+?>
+
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Title</title>
+</head>
+<body>
+
+
+
+
+</body>
+</html>
