--- web/presentation/controllers/dashboard.php
@@ -1,21 +1,0 @@
-<?php
-
-include '../security/session_check.php';
-
-
-require_once '../../includes/models/LoginRequest.php';
-require_once '../../includes/models/Token.php';
-
-use models\LoginRequest;
-use models\Token;
-
-
-if (!isset($_SESSION['SESSION_TOKEN'])) {
-    header("Location: ../views/login/login.html");
-    exit();
-}
-
-
-$jwtToken = Token::fromJson($_SESSION['SESSION_TOKEN']);
-
-
