--- web/presentation/controllers/logout.php
@@ -1,23 +1,0 @@
-<?php
-session_start();
-
-// Unset all session variables
-$_SESSION = [];
-
-// Destroy the session
-session_destroy();
-
-// Clear session cookies
-if (isset($_COOKIE[session_name()])) {
-    setcookie(session_name(), '', time() - 3600, '/');
-}
-
-// Prevent accessing pages after logout by disabling caching
-header("Cache-Control: no-store, no-cache, must-revalidate, max-age=0");
-header("Pragma: no-cache");
-header("Expires: Sat, 01 Jan 2000 00:00:00 GMT");
-
-// Redirect to login page
-header("Location: ../views/login/login.html?error=Logged Out");
-exit();
-
