--- web/presentation/controllers/queue.php
@@ -1,3 +1,0 @@
-<?php
-
-
