--- web/presentation/views/login/login.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Admin Login</title>
-    <link rel="stylesheet" href="../../../css/styles.css">
-</head>
-<body>
-<div class="login-container">
-    <img src="../../../static/images/University_of_Pangasinan_logo%20(1).png" alt="Logo" class="logo">
-    <h2>UPang Registrar</h2>
-    <form method="POST" action="../../controllers/login.php">
-        <label>Username</label>
-        <input type="text" name="username" placeholder="staff-id" required>
-        <label>Password</label>
-        <input type="password" name="password" placeholder="Password" required>
-        <p id="error-message"></p> <!-- Error Message Area -->
-        <button type="submit">Log in</button>
-    </form>
-    <a href="#" class="forgot-password">Forgot Password?</a>
-</div>
-
-<script type="module">
-    import { showErrorFromURL } from "../../../js/errorHandler.js";
-    showErrorFromURL();
-</script>
-</body>
-</html>
