--- web/presentation/views/queue/queue.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Queue Updates</title>
-    <link rel="stylesheet" href="../../../css/dashboard.css">
-</head>
-<body>
-
-<!-- Navigation -->
-<div class="navbar">
-    <div class="logo-container">
-        <img src="../../../static/images/University_of_Pangasinan_logo%20(1).png" alt="Logo" class="logo">
-        <h2>UNIVERSITY OF PANGASINAN</h2>
-    </div>
-    <nav>
-        <a href="../dashboard/dashboard.php">Back to Dashboard</a>
-    </nav>
-</div>
-
-<!-- Content Section -->
-<div class="content">
-    <h3>Queue Updates</h3>
-
-
-    <div class="queue-box">
-        <?php
-            include_once '../../controllers/queue.php';
-        ?>
-    </div>
-</div>
-
-<footer class="footer">
-    <p>&copy;2025 Upang-Registrar. All rights reserved</p>
-</footer>
-
-</body>
-</html>
