+++ web/public/js/errorHandler.js
