+++ web/public/logout.php
@@ -1,0 +1,1 @@
+<?php
